refactor(cli): rename ck skill to ck skills
Rename CLI command from singular to plural to match industry standard
(npx add-skill, Agent Skills spec).

- Directory: src/commands/skill/ → src/commands/skills/
- Command: ck skill → ck skills
- No backward compat alias (clean break)
- Registry path unchanged

Closes #336
