+++ src/cli/command-registry.ts
@@ -9,7 +9,7 @@ import { doctorCommand } from "../commands/doctor.js";
 import { easterEggCommand } from "../commands/easter-egg.js";
 import { initCommand } from "../commands/init.js";
 import { newCommand } from "../commands/new/index.js";
-import { skillCommand } from "../commands/skill/index.js";
+import { skillsCommand } from "../commands/skills/index.js";
 import { uninstallCommand } from "../commands/uninstall/index.js";
 import { updateCliCommand } from "../commands/update-cli.js";
 import { versionCommand } from "../commands/version.js";
@@ -200,9 +200,9 @@ export function registerCommands(cli: ReturnType<typeof cac>): void {
 			await easterEggCommand();
 		});
 
-	// Skill command - install skills to other coding agents
+	// Skills command - install skills to other coding agents
 	cli
-		.command("skill", "Install ClaudeKit skills to other coding agents")
+		.command("skills", "Install ClaudeKit skills to other coding agents")
 		.option("-n, --name <skill>", "Skill name to install/uninstall")
 		.option("-a, --agent <agents...>", "Target agents (claude-code, cursor, codex, etc.)")
 		.option("-g, --global", "Install/uninstall globally instead of project-level")
@@ -218,6 +218,6 @@ export function registerCommands(cli: ReturnType<typeof cac>): void {
 			if (options.agent && !Array.isArray(options.agent)) {
 				options.agent = [options.agent];
 			}
-			await skillCommand(options);
+			await skillsCommand(options);
 		});
 }
