+++ src/commands/skills/__tests__/agents.test.ts
