+++ src/commands/skills/agents.ts
