+++ src/commands/skills/skills-discovery.ts
