+++ src/commands/skills/skills-registry.ts
