+++ src/domains/help/commands/index.ts
@@ -10,7 +10,7 @@ export { doctorCommandHelp } from "./doctor-command-help.js";
 export { uninstallCommandHelp } from "./uninstall-command-help.js";
 export { updateCommandHelp } from "./update-command-help.js";
 export { versionsCommandHelp } from "./versions-command-help.js";
-export { skillCommandHelp } from "./skill-command-help.js";
+export { skillsCommandHelp } from "./skills-command-help.js";
 export {
 	projectOptionsGroup,
 	filterOptionsGroup,
